add user-controlled body parser and cookie configs
we think the reason apple webhook events are not received has to do with this middleware setup. There are two issues that are fixed here.

One is that no middleware should be applied before sentry and logging middleware. Configuring body parser and cookies _after_ the user's custom boot config is brought in, allowing them to wrap their own logger/exception handlers prior to any middleware being setup

The other is that we need a way to expoee post body size limits to the user through a config, so that it can shift app to app based on need.

related https://rvohealth.atlassian.net/browse/PDTC-3288