add Params.expect
Params#expect is a method useful for type narrowing. A need for this
cropped up after we applied typescript-specific eslint rules to our
project. Now that any is unsafe, we need a way to both restrict a type,
and at the same type intelligently send back a return type, iff the param
field matches the type you are looking for.

close https://rvohealth.atlassian.net/browse/PDTC-5120