boost performance for send spec helper
although it seems a fairly minor change to cache a variable which is
already being cached by the underlying function, the cost of waiting
for the promise to return, even with a simple caching layer on the other
side of the function being consumed, is still high enough to warrant
speeding up. Testing this fix in a particularly request-heavy test
(it had 32 individual request specs in it) within wellos, the test went from running
in `37 seconds` to running in `47 seconds`. Adding this fix sped the
test back up to 37 seconds again.