Remove callback function approach to passing
classes to @OpenAPI Decorator. Passing in a callback
function was only to be consistent with how association
decorators worked in models, where the callback functions
prevented the problem of circular class references during
file loading. That approach recently broke when one of our
projects reached a certain level of complexity, so we've
changed models to referencing one another and serializers
by string.