add PsychicDevtools export in replacement of FrontEndClient
FrontEndClient was short-sighted. We want psychic to be flexible, but
this pattern encouraged a single-client modality, in which psychic was
responsible for launching the server. We are deprecating this pattern
and moving towards a psychic conf that leverages PsychicDevtools to
launch a server, and then uses it again on shut down to shut it down.

Since this introduces a minor breaking change to devs, I am bumping the
minor. Almost to v1!!!