remove openapi-express-validator
This library is incredibly easy to set up if you want to use it, but to
reduce risk of exposure to vulnerable packages, we are removing direct
integration with it. Instead, I will replace with a tutorial in our
psychic-guides repo on how to utilize this library via `conf/app.ts`
hooks if one desires.

It may also be nice to:

1.) add to boilerplate?
2.) (if not 1) add a new setup cli command for it