remove package manager verification
a.) doing a package manager check this way makes it difficult for us to
test out new package managers in create-psychic, since it will prevent
unrecognized package managers from coming through.
b.) this check should be in dream, since both psychic and dream depend
on the packageManager arg. When it moves to dream, it should be done in
a way that can be bypassed by env var, so that create-psychic PRs can
use it to temporarily bypass the check.